change api->include and hapi->api
