+++ paddle/fluid/operators/scale_op.h
@@ -18,9 +18,9 @@ limitations under the License. */
 #include "paddle/fluid/framework/pten_utils.h"
 
 // only can include the headers in paddle/top/api dirs
-#include "paddle/pten/api/include/core.h"
-#include "paddle/pten/api/include/math.h"
-#include "paddle/pten/hapi/lib/utils/tensor_utils.h"
+#include "paddle/pten/api/lib/utils/tensor_utils.h"
+#include "paddle/pten/include/core.h"
+#include "paddle/pten/include/math.h"
 
 namespace paddle {
 namespace operators {
