+++ paddle/pten/CMakeLists.txt
@@ -1,7 +1,6 @@
-# pten api
+# pten (low level) api headers: include
+# pten (high level) api
 add_subdirectory(api)
-# pten high level api
-add_subdirectory(hapi)
 # pten core components
 add_subdirectory(core)
 # pten kernels for diff device
@@ -10,3 +9,13 @@ add_subdirectory(kernels)
 add_subdirectory(infershape)
 # pten tests
 add_subdirectory(tests)
+
+# make an unity target for compile deps
+set(PTEN_DEPS convert_utils dense_tensor kernel_factory kernel_context)
+set(PTEN_DEPS ${PTEN_DEPS} math_cpu linalg_cpu creation_cpu manipulation_cpu)
+set(PTEN_DEPS ${PTEN_DEPS} unary binary)
+if(WITH_GPU OR WITH_ROCM)
+  set(PTEN_DEPS ${PTEN_DEPS} math_cuda linalg_cuda creation_cuda manipulation_cuda)
+endif()
+
+cc_library(pten SRCS all.cc DEPS ${PTEN_DEPS})
