+++ paddle/pten/api/CMakeLists.txt
@@ -1,8 +1,3 @@
-set(PTEN_DEPS convert_utils dense_tensor kernel_factory kernel_context)
-set(PTEN_DEPS ${PTEN_DEPS} math_cpu linalg_cpu creation_cpu manipulation_cpu)
-set(PTEN_DEPS ${PTEN_DEPS} unary binary)
-if(WITH_GPU OR WITH_ROCM)
-  set(PTEN_DEPS ${PTEN_DEPS} math_cuda linalg_cuda creation_cuda manipulation_cuda)
-endif()
+add_subdirectory(lib)
 
-cc_library(pten SRCS all.cc DEPS ${PTEN_DEPS})
+cc_library(pten_hapi SRCS all.cc DEPS linalg_api math_api creation_api)
