+++ paddle/pten/api/include/creation.h
@@ -14,26 +14,20 @@
 
 #pragma once
 
-#include "paddle/pten/api/include/infershape.h"
-#include "paddle/pten/hapi/lib/utils/allocator.h"
-#include "paddle/pten/kernels/cpu/creation.h"
-#include "paddle/pten/kernels/cuda/creation.h"
+#include "paddle/pten/api/include/tensor.h"
+#include "paddle/pten/common/data_type.h"
+#include "paddle/pten/common/scalar.h"
 
-namespace pten {
+namespace paddle {
+namespace experimental {
 
-// TODO(YuanRisheng) This function name should be same as User API name.
-// TODO(zyfncg) Automatic code generation
-template <typename T, typename ContextT>
-DenseTensor FillAnyLike(const ContextT& dev_ctx,
-                        const DenseTensor& x,
-                        const Scalar& val) {
-  auto out_meta = UnchangedInferShape(x.meta());
-  const auto allocator =
-      std::make_shared<paddle::experimental::DefaultAllocator>(
-          dev_ctx.GetPlace());
-  pten::DenseTensor dense_out(allocator, out_meta);
-  FillAnyLike<T>(dev_ctx, x, val, &dense_out);
-  return dense_out;
-}
+Tensor full_like(const Tensor& x,
+                 const Scalar& value,
+                 DataType dtype = DataType::UNDEFINED);
 
-}  // namespace pten
+Tensor ones_like(const Tensor& x, DataType dtype = DataType::UNDEFINED);
+
+Tensor zeros_like(const Tensor& x, DataType dtype = DataType::UNDEFINED);
+
+}  // namespace experimental
+}  // namespace paddle
