+++ paddle/pten/api/include/linalg.h
@@ -14,25 +14,17 @@
 
 #pragma once
 
-// See Note: [ How do we organize the kernel directory ]
-#include "paddle/pten/api/include/infershape.h"
-#include "paddle/pten/hapi/lib/utils/allocator.h"
-#include "paddle/pten/kernels/cpu/linalg.h"
-#include "paddle/pten/kernels/cuda/linalg.h"
+#include "paddle/pten/api/include/tensor.h"
 
-namespace pten {
+namespace paddle {
+namespace experimental {
 
-template <typename T, typename ContextT>
-DenseTensor Dot(const ContextT& dev_ctx,
-                const DenseTensor& x,
-                const DenseTensor& y) {
-  auto out_meta = DotInferShape(x.meta(), y.meta());
-  const auto allocator =
-      std::make_shared<paddle::experimental::DefaultAllocator>(
-          dev_ctx.GetPlace());
-  pten::DenseTensor dense_out(allocator, out_meta);
-  Dot<T>(dev_ctx, x, y, &dense_out);
-  return dense_out;
-}
+Tensor dot(const Tensor& x, const Tensor& y);
 
-}  // namespace pten
+Tensor matmul(const Tensor& x,
+              const Tensor& y,
+              bool transpose_x,
+              bool transpose_y);
+
+}  // namespace experimental
+}  // namespace paddle
