+++ paddle/pten/api/lib/linalg.cc
@@ -12,19 +12,19 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License. */
 
-#include "paddle/pten/hapi/include/linalg.h"
+#include "paddle/pten/api/include/linalg.h"
 
 #include <memory>
 
 #include "glog/logging.h"
 
-#include "paddle/pten/api/include/core.h"
-#include "paddle/pten/api/include/infershape.h"
+#include "paddle/pten/api/lib/kernel_dispatch.h"
+#include "paddle/pten/api/lib/utils/allocator.h"
 #include "paddle/pten/core/convert_utils.h"
 #include "paddle/pten/core/dense_tensor.h"
 #include "paddle/pten/core/kernel_context.h"
-#include "paddle/pten/hapi/lib/kernel_dispatch.h"
-#include "paddle/pten/hapi/lib/utils/allocator.h"
+#include "paddle/pten/include/core.h"
+#include "paddle/pten/include/infershape.h"
 
 namespace paddle {
 namespace experimental {
