--- paddle/pten/hapi/include/math.h
@@ -1,27 +1,0 @@
-/* Copyright (c) 2021 PaddlePaddle Authors. All Rights Reserved.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License. */
-
-#pragma once
-
-#include "paddle/pten/hapi/include/tensor.h"
-
-namespace paddle {
-namespace experimental {
-
-// TODO(chenweihang): add scale API
-// TODO(chenweihang): move mean API into stat.h/cc
-Tensor mean(const Tensor& x);
-
-}  // namespace experimental
-}  // namespace paddle
