+++ paddle/pten/tests/test_copy_api.cc
@@ -18,8 +18,8 @@ limitations under the License. */
 #include "paddle/pten/core/kernel_registry.h"
 #include "paddle/pten/kernels/cpu/utils.h"
 
+#include "paddle/pten/api/lib/utils/allocator.h"
 #include "paddle/pten/core/dense_tensor.h"
-#include "paddle/pten/hapi/lib/utils/allocator.h"
 
 PT_DECLARE_MODULE(UtilsCPU);
